feat: update pricing to $9.99/mo Pro, $19.99/mo Team with 40% annual discount
Pro: $9.99/month, $5.99/month billed annually ($71.88/year)
Team: $19.99/month, $11.99/month billed annually ($143.88/year)
Updated fallback defaults, Schema.org, admin MRR calculation,
translation strings, and parseFloat for decimal price support.

+++ lib/subscription-utils.ts
@@ -31,8 +31,8 @@ export function buildTier(prefix: string, config: Record<string, string>): TierD
     receiptScanLimit: parseInt(config[`${prefix}_receipt_scan_limit`] || '0'),
     emailSendLimit: parseInt(config[`${prefix}_email_send_limit`] || '0'),
     storageMb: parseInt(config[`${prefix}_storage_mb`] || '0'),
-    priceMonthly: parseInt(config[`${prefix}_price_monthly`] || '0'),
-    priceYearly: parseInt(config[`${prefix}_price_yearly`] || '0'),
+    priceMonthly: parseFloat(config[`${prefix}_price_monthly`] || '0'),
+    priceYearly: parseFloat(config[`${prefix}_price_yearly`] || '0'),
     features: parseJsonArray(config[`${prefix}_features`], []),
   }
 }
